APP-4338: Update `TextArea` minimum height when `rows` attribute is provided
The HTML `rows` attribute defines the number of initial lines to display in the HTML `textarea`. When given, it enters in conflict with the `min-height` of the `TextArea`.
Today, the `min-height` is set to `110px` (~5 lines), so when the given `rows` is greater or equal to `5`, it works, but not when it's lower than `5` (because the textarea will have a minimum of 5 lines).

The solution here is to set the `min-height` only when the `rows` is not given.