APP-4916: Add CreatableSelect and CreatableAsyncSeelct to the Dropdown component
This commit adds a new isCreatable prop to the Dropdown component.
When this prop is true, the user can create new options at runtime, according to the inputValue. This is compatible with simple selects, multi select and async selects.

Optionally, 2 other props can be provided along with this isCreatable prop:
- getNewOptionData: function telling how to structure the created option based on the user input (default one will create options using this structure { value: , label:  })
- isValidNewOption: function telling if the inputValue can be converted into a new option (default one always returns true)

Briefly:
- the DropdownTag is now chosen according to the isCreatable and asyncOptions props (see Dropdown.tsx)
- adds related interfaces (see CreatableProps<T> in interfaces.ts)
- adds related unit tests (see Dropdown.spec.tsx)
- adds related stories (see Dropdown.stories.tsx)